Project Start
Initial state of the project. The project contains a timer object that
can be observed by any other object that implements an update method.
Upon every second the timer will notify all observers that the state of
the timer has changed, and then the observers can query the timer for
the data they need. A test webpage is included to test functionality of
the timer before it becomes an extension.