Major Project Overhaul
Many things were done in this commit. The project was restructured to
accommodate using gulp.js to minify and move files. JS files are now
minified into a new distribution folder on the same level as src (which
is not being included in the repo). The gulp file and package.json to
specify gulp dependencies is now included in the repo to allow
development from any machine.

The custom library JS files also were overhauled to support the use of a
namespace. This will make the objects from the library accessible to
other scripts outside of the background.js that may need to use the
objects.